requirements.txt
# Dudley Life Optimizer - Python Dependencies

# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Database & ORM
supabase==2.3.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# AI & Machine Learning
openai==1.6.1
anthropic==0.8.1
numpy==1.24.3
scikit-learn==1.3.2

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.8
pyotp==2.9.0
qrcode[pil]==7.4.2

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3

# HTTP & API
requests==2.31.0
urllib3==2.1.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.2

# Logging & Monitoring
structlog==23.2.0

# Image Processing
Pillow==10.1.0

# Validation
marshmallow==3.20.2

# Async Support
asyncio==3.4.3
aiohttp==3.9.1

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Utilities
click==8.1.7
python-slugify==8.0.1
validators==0.22.0

# Vector Database
pgvector==0.2.4

# Time Series
arrow==1.3.0

# Caching
redis==5.0.1

# File Handling
openpyxl==3.1.2
python-magic==0.4.27

# Network & Security
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

# Math & Statistics
scipy==1.11.4
statsmodels==0.14.1

# Visualization (for analytics)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2

# Health & Fitness APIs
fitbit==0.3.1
polar-accesslink-python==1.0.0

# Business Intelligence
yfinance==0.2.28

# Email & Notifications
sendgrid==6.11.0
twilio==8.12.0

# Scheduling
APScheduler==3.10.4

# Configuration Management
dynaconf==3.2.4

# API Documentation
flask-restx==1.3.0
flasgger==0.9.7.1

# Performance Monitoring
prometheus-client==0.19.0

# Data Serialization
msgpack==1.0.7

# Timezone Handling
babel==2.14.0

# UUID Generation
uuid==1.30

# Regular Expressions
regex==2023.10.3

# JSON Handling
orjson==3.9.10

# Encryption
fernet==1.0.1

# Rate Limiting
slowapi==0.1.9

# Health Checks
healthcheck==1.3.3

# Memory Profiling
memory-profiler==0.61.0

# Process Management
psutil==5.9.6

# Backup & Archive
zipfile36==0.1.3

# Data Validation
cerberus==1.3.5

# API Rate Limiting
limits==3.7.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Monitoring
sentry-sdk[flask]==1.39.2

# Development Tools
ipython==8.18.1
jupyter==1.0.0

# Code Quality
pre-commit==3.6.0
bandit==1.7.5

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

